requirements.txt
aiohttp==3.9.3
aiolimiter==1.1.0
aiosignal==1.3.1
anthropic==0.19.2
anyio==3.7.1
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==4.45
coloredlogs==15.0.1
dataclasses-json==0.6.4
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fake-useragent==1.5.1
fastapi==0.110.3
fastavro==1.9.4
filelock==3.13.1
fireworks-ai==0.13.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.60.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.4.0
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.12
langchain-anthropic==0.1.4
langchain-community==0.0.28
langchain-core==0.1.31
langchain-fireworks==0.1.1
langchain-google-genai==0.0.6
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langserve==0.0.39
langsmith==0.1.51
lxml==4.9.4
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.8
onnxruntime==1.17.3
openai==1.12.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pillow==10.2.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pulsar-client==3.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.0
Pygments==2.17.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruff==0.2.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sse-starlette==1.8.2
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.2
tqdm==4.66.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.19.0
voyageai==0.1.7
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
